Refactored navigation and routing
