+++ src/features/visualizations/covid/racing/RacingBarChart.js
